Integradora/requirements.txt
Flask
mysql.connector 